20211113 data transfer test from server to client
