+++ DB_server/models/users.js
@@ -4,12 +4,8 @@ module.exports = class Users extends Sequelize.Model {
   static init(sequelize) {
     return super.init(
       {
-        id: {
-          primaryKey: true,
-          type: Sequelize.STRING(45),
-          allowNull: false,
-        },
         email: {
+          primaryKey: true,
           type: Sequelize.STRING(45),
           allowNull: false,
         },
@@ -35,9 +31,13 @@ module.exports = class Users extends Sequelize.Model {
 
   // 테이블간 관계 설정
   static associate(db) {
-    // db.Users.hasMany(db.Assets, {
-    //   foreignKey: "email",
-    //   sourceKey: "email",
-    // });
+    db.Users.hasMany(db.Assets, {
+      foreignKey: "email",
+      sourceKey: "email",
+    });
+    db.Users.hasMany(db.Comments, {
+      foreignKey: "email",
+      sourceKey: "email",
+    });
   }
 };
