+++ DB_server/server.js
@@ -4,7 +4,7 @@ const { sequelize } = require("./models/index.js");
 
 const app = express();
 
-const indexRouter = require("./routes/index.js");
+const indexRouter = require("./routers/index.js");
 
 // 포트 설정
 const PORT = 7000;
@@ -26,7 +26,7 @@ app.use((err, req, res, next) => {
 sequelize
   // sync : MySQL에 테이블이 존재 하지 않을때 생성
   //      force: true   => 이미 테이블이 있으면 drop하고 다시 테이블 생성
-  .sync({ force: false })
+  .sync({ force: true })
   .then(() => {
     console.log("Database connected successfully");
   })
